Cleanup after self-review
