+++ pkg/testing/tools/fleettools/fleet.go
@@ -206,104 +206,3 @@ func UpdateESOutputPreset(ctx context.Context, client *kibana.Client, outputID s
 
 	return nil
 }
-
-// osqueryAPIVersion is the versioned-route header required by the osquery
-// plugin's public API. See
-// x-pack/platform/plugins/shared/osquery/common/constants.ts (API_VERSIONS.public.v1)
-// in the Kibana repository.
-const osqueryAPIVersion = "2023-10-31"
-
-func osqueryAPIHeaders() http.Header {
-	h := http.Header{}
-	h.Set("elastic-api-version", osqueryAPIVersion)
-	return h
-}
-
-// doOsqueryRequest sends a request to the osquery plugin's API and returns the
-// raw response body, after checking for a 200 status. action names the call
-// for error messages (e.g. "submitting osquery live query").
-func doOsqueryRequest(ctx context.Context, client *kibana.Client, method, path, action string, reqBody io.Reader) ([]byte, error) {
-	resp, err := client.SendWithContext(ctx, method, path, nil, osqueryAPIHeaders(), reqBody)
-	if err != nil {
-		return nil, fmt.Errorf("%s: %w", action, err)
-	}
-	defer resp.Body.Close()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("reading %s response: %w", action, err)
-	}
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("%s returned status %d: %s", action, resp.StatusCode, body)
-	}
-	return body, nil
-}
-
-// OsqueryLiveQuery identifies a submitted live query: ActionID is the parent
-// action ID (used to poll status), QueryActionID is the per-query action ID
-// (used to fetch results).
-type OsqueryLiveQuery struct {
-	ActionID      string
-	QueryActionID string
-}
-
-type osqueryLiveQueryCreateResponse struct {
-	Data struct {
-		ActionID string `json:"action_id"`
-		Queries  []struct {
-			ActionID string `json:"action_id"`
-		} `json:"queries"`
-	} `json:"data"`
-}
-
-// SubmitOsqueryLiveQuery submits a live query for the given agent via Kibana's
-// osquery plugin (POST /api/osquery/live_queries), which creates a Fleet
-// INPUT_ACTION with input_type "osquery" under the hood.
-func SubmitOsqueryLiveQuery(ctx context.Context, client *kibana.Client, agentID, query string) (OsqueryLiveQuery, error) {
-	reqBody, err := json.Marshal(map[string]any{
-		"agent_ids": []string{agentID},
-		"query":     query,
-	})
-	if err != nil {
-		return OsqueryLiveQuery{}, fmt.Errorf("marshaling live query request: %w", err)
-	}
-
-	body, err := doOsqueryRequest(ctx, client, http.MethodPost, "/api/osquery/live_queries", "submitting osquery live query", bytes.NewReader(reqBody))
-	if err != nil {
-		return OsqueryLiveQuery{}, err
-	}
-
-	var parsed osqueryLiveQueryCreateResponse
-	if err := json.Unmarshal(body, &parsed); err != nil {
-		return OsqueryLiveQuery{}, fmt.Errorf("unmarshaling osquery live query response: %w: %s", err, body)
-	}
-	if parsed.Data.ActionID == "" || len(parsed.Data.Queries) == 0 {
-		return OsqueryLiveQuery{}, fmt.Errorf("osquery live query response missing action id(s): %s", body)
-	}
-
-	return OsqueryLiveQuery{
-		ActionID:      parsed.Data.ActionID,
-		QueryActionID: parsed.Data.Queries[0].ActionID,
-	}, nil
-}
-
-type osqueryLiveQueryDetailsResponse struct {
-	Data struct {
-		Status string `json:"status"`
-	} `json:"data"`
-}
-
-// GetOsqueryLiveQueryStatus fetches the status ("running" or "completed") of a
-// previously submitted live query via GET /api/osquery/live_queries/{actionID}.
-func GetOsqueryLiveQueryStatus(ctx context.Context, client *kibana.Client, actionID string) (string, error) {
-	body, err := doOsqueryRequest(ctx, client, http.MethodGet, "/api/osquery/live_queries/"+actionID, "fetching osquery live query details", nil)
-	if err != nil {
-		return "", err
-	}
-
-	var parsed osqueryLiveQueryDetailsResponse
-	if err := json.Unmarshal(body, &parsed); err != nil {
-		return "", fmt.Errorf("unmarshaling osquery live query details response: %w: %s", err, body)
-	}
-	return parsed.Data.Status, nil
-}
